Tests/LibWeb/Text/expected/Crypto/SubtleCrypto-structuredClone.txt

clone.type=secret extractable=false
clone.algorithm.name=PBKDF2
clone.usages=["deriveBits","deriveKey"]
derived.algorithm.name=AES-CBC
